+++ src/main/java/org/opensearch/neuralsearch/common/VectorUtil.java
@@ -2,7 +2,6 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-
 package org.opensearch.neuralsearch.common;
 
 import java.util.List;