+++ src/main/java/org/opensearch/neuralsearch/query/visitor/NeuralSearchQueryVisitor.java
@@ -2,7 +2,6 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-
 package org.opensearch.neuralsearch.query.visitor;
 
 import java.util.Map;