+++ src/test/java/org/opensearch/neuralsearch/query/HybridQueryBuilderTests.java
@@ -2,7 +2,6 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-
 package org.opensearch.neuralsearch.query;
 
 import static org.hamcrest.Matchers.containsString;