+++ src/test/java/org/opensearch/neuralsearch/query/HybridQueryIT.java
@@ -2,13 +2,15 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-
 package org.opensearch.neuralsearch.query;
 
 import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.containsString;
 import static org.opensearch.index.query.QueryBuilders.matchQuery;
 import static org.opensearch.neuralsearch.TestUtils.DELTA_FOR_SCORE_ASSERTION;
+import static org.opensearch.neuralsearch.TestUtils.RELATION_EQUAL_TO;
+import static org.opensearch.neuralsearch.TestUtils.TEST_DIMENSION;
+import static org.opensearch.neuralsearch.TestUtils.TEST_SPACE_TYPE;
 import static org.opensearch.neuralsearch.TestUtils.createRandomVector;
 
 import java.io.IOException;
@@ -29,8 +31,7 @@ import org.opensearch.index.query.MatchQueryBuilder;
 import org.opensearch.index.query.NestedQueryBuilder;
 import org.opensearch.index.query.QueryBuilders;
 import org.opensearch.index.query.TermQueryBuilder;
-import org.opensearch.knn.index.SpaceType;
-import org.opensearch.neuralsearch.common.BaseNeuralSearchIT;
+import org.opensearch.neuralsearch.BaseNeuralSearchIT;
 
 import com.google.common.primitives.Floats;
 
@@ -55,9 +56,6 @@ public class HybridQueryIT extends BaseNeuralSearchIT {
     private static final String TEST_KNN_VECTOR_FIELD_NAME_2 = "test-knn-vector-2";
     private static final String TEST_TEXT_FIELD_NAME_1 = "test-text-field-1";
     private static final String TEST_NESTED_TYPE_FIELD_NAME_1 = "user";
-
-    private static final int TEST_DIMENSION = 768;
-    private static final SpaceType TEST_SPACE_TYPE = SpaceType.L2;
     private static final String NESTED_FIELD_1 = "firstname";
     private static final String NESTED_FIELD_2 = "lastname";
     private static final String NESTED_FIELD_1_VALUE = "john";
@@ -65,7 +63,6 @@ public class HybridQueryIT extends BaseNeuralSearchIT {
     private final float[] testVector1 = createRandomVector(TEST_DIMENSION);
     private final float[] testVector2 = createRandomVector(TEST_DIMENSION);
     private final float[] testVector3 = createRandomVector(TEST_DIMENSION);
-    private final static String RELATION_EQUAL_TO = "eq";
     private static final String SEARCH_PIPELINE = "phase-results-pipeline";
 
     @Before