+++ src/testFixtures/java/org/opensearch/neuralsearch/BaseNeuralSearchIT.java
@@ -2,8 +2,7 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-
-package org.opensearch.neuralsearch.common;
+package org.opensearch.neuralsearch;
 
 import static org.opensearch.neuralsearch.common.VectorUtil.vectorAsListToArray;
 
@@ -26,6 +25,7 @@ import org.apache.hc.core5.http.ContentType;
 import org.apache.hc.core5.http.Header;
 import org.apache.hc.core5.http.HttpEntity;
 import org.apache.hc.core5.http.HttpHeaders;
+import org.apache.hc.core5.http.ParseException;
 import org.apache.hc.core5.http.io.entity.EntityUtils;
 import org.apache.hc.core5.http.io.entity.StringEntity;
 import org.apache.hc.core5.http.message.BasicHeader;
@@ -45,7 +45,6 @@ import org.opensearch.core.xcontent.ToXContent;
 import org.opensearch.core.xcontent.XContentBuilder;
 import org.opensearch.index.query.QueryBuilder;
 import org.opensearch.knn.index.SpaceType;
-import org.opensearch.neuralsearch.OpenSearchSecureRestTestCase;
 import org.opensearch.neuralsearch.util.NeuralSearchClusterUtil;
 import org.opensearch.test.ClusterServiceUtils;
 import org.opensearch.threadpool.TestThreadPool;
@@ -135,11 +134,13 @@ public abstract class BaseNeuralSearchIT extends OpenSearchSecureRestTestCase {
         assertEquals(RestStatus.OK, RestStatus.fromCode(response.getStatusLine().getStatusCode()));
     }
 
-    protected String uploadModel(String requestBody) throws Exception {
-        String modelGroupId = registerModelGroup();
+    protected String registerModelGroupAndUploadModel(String requestBody) throws Exception {
+        String modelGroupId = getModelGroupId();
         // model group id is dynamically generated, we need to update model update request body after group is registered
-        requestBody = requestBody.replace("<MODEL_GROUP_ID>", modelGroupId);
+        return uploadModel(String.format(LOCALE, requestBody, modelGroupId));
+    }
 
+    protected String uploadModel(String requestBody) throws Exception {
         Response uploadResponse = makeRequest(
             client(),
             "POST",
@@ -202,7 +203,7 @@ public abstract class BaseNeuralSearchIT extends OpenSearchSecureRestTestCase {
     @SneakyThrows
     protected String prepareModel() {
         String requestBody = Files.readString(Path.of(classLoader.getResource("processor/UploadModelRequestBody.json").toURI()));
-        String modelId = uploadModel(requestBody);
+        String modelId = registerModelGroupAndUploadModel(requestBody);
         loadModel(modelId);
         return modelId;
     }
@@ -270,18 +271,17 @@ public abstract class BaseNeuralSearchIT extends OpenSearchSecureRestTestCase {
     }
 
     protected void createPipelineProcessor(String modelId, String pipelineName, ProcessorType processorType) throws Exception {
+        String requestBody = Files.readString(Path.of(classLoader.getResource(PIPELINE_CONFIGS_BY_TYPE.get(processorType)).toURI()));
+        createPipelineProcessor(requestBody, pipelineName, modelId);
+    }
+
+    protected void createPipelineProcessor(String requestBody, String pipelineName, String modelId) throws Exception {
         Response pipelineCreateResponse = makeRequest(
             client(),
             "PUT",
             "/_ingest/pipeline/" + pipelineName,
             null,
-            toHttpEntity(
-                String.format(
-                    LOCALE,
-                    Files.readString(Path.of(classLoader.getResource(PIPELINE_CONFIGS_BY_TYPE.get(processorType)).toURI())),
-                    modelId
-                )
-            ),
+            toHttpEntity(String.format(LOCALE, requestBody, modelId)),
             ImmutableList.of(new BasicHeader(HttpHeaders.USER_AGENT, DEFAULT_USER_AGENT))
         );
         Map<String, Object> node = XContentHelper.convertToMap(
@@ -790,10 +790,16 @@ public abstract class BaseNeuralSearchIT extends OpenSearchSecureRestTestCase {
     }
 
     @SneakyThrows
-    private String registerModelGroup() {
+    private String getModelGroupId() {
         String modelGroupRegisterRequestBody = Files.readString(
             Path.of(classLoader.getResource("processor/CreateModelGroupRequestBody.json").toURI())
-        ).replace("<MODEL_GROUP_NAME>", "public_model_" + RandomizedTest.randomAsciiAlphanumOfLength(8));
+        );
+        return registerModelGroup(
+            String.format(LOCALE, modelGroupRegisterRequestBody, "public_model_" + RandomizedTest.randomAsciiAlphanumOfLength(8))
+        );
+    }
+
+    protected String registerModelGroup(String modelGroupRegisterRequestBody) throws IOException, ParseException {
         Response modelGroupResponse = makeRequest(
             client(),
             "POST",
@@ -810,6 +816,63 @@ public abstract class BaseNeuralSearchIT extends OpenSearchSecureRestTestCase {
         String modelGroupId = modelGroupResJson.get("model_group_id").toString();
         assertNotNull(modelGroupId);
         return modelGroupId;
+    }
+
+    // Method that waits till the health of nodes in the cluster goes green
+    protected void waitForClusterHealthGreen(String numOfNodes) throws IOException {
+        Request waitForGreen = new Request("GET", "/_cluster/health");
+        waitForGreen.addParameter("wait_for_nodes", numOfNodes);
+        waitForGreen.addParameter("wait_for_status", "green");
+        client().performRequest(waitForGreen);
+    }
+
+    /**
+     * Add a single Doc to an index
+     *
+     * @param index name of the index
+     * @param docId
+     * @param fieldName name of the field
+     * @param text to be added
+     */
+    protected void addDocument(String index, String docId, String fieldName, String text) throws IOException {
+        Request request = new Request("PUT", "/" + index + "/_doc/" + docId + "?refresh=true");
+
+        XContentBuilder builder = XContentFactory.jsonBuilder().startObject().field(fieldName, text).endObject();
+        request.setJsonEntity(builder.toString());
+        client().performRequest(request);
+    }
+
+    /**
+     * Get ingest pipeline
+     * @param pipelineName of the ingest pipeline
+     *
+     * @return get pipeline response as a map object
+    */
+    @SneakyThrows
+    protected Map<String, Object> getIngestionPipeline(String pipelineName) {
+        Request request = new Request("GET", "/_ingest/pipeline/" + pipelineName);
+        Response response = client().performRequest(request);
+        assertEquals(request.getEndpoint() + ": failed", RestStatus.OK, RestStatus.fromCode(response.getStatusLine().getStatusCode()));
+        String responseBody = EntityUtils.toString(response.getEntity());
+        Map<String, Object> responseMap = createParser(XContentType.JSON.xContent(), responseBody).map();
+        return (Map<String, Object>) responseMap.get(pipelineName);
+    }
+
+    /**
+     * Delete pipeline
+     *
+     * @param pipelineName of the pipeline
+     *
+     * @return delete pipeline response as a map object
+     */
+    @SneakyThrows
+    protected Map<String, Object> deletePipeline(String pipelineName) {
+        Request request = new Request("DELETE", "/_ingest/pipeline/" + pipelineName);
+        Response response = client().performRequest(request);
+        assertEquals(request.getEndpoint() + ": failed", RestStatus.OK, RestStatus.fromCode(response.getStatusLine().getStatusCode()));
+        String responseBody = EntityUtils.toString(response.getEntity());
+        Map<String, Object> responseMap = createParser(XContentType.JSON.xContent(), responseBody).map();
+        return responseMap;
     }
 
     /**