+++ CHANGELOG.md
@@ -22,6 +22,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Bug Fixes
 - Fixing multiple issues reported in #497 ([#524](https://github.com/opensearch-project/neural-search/pull/524))
 ### Infrastructure
+- BWC tests for Neural Search ([#515](https://github.com/opensearch-project/neural-search/pull/515))
 ### Documentation
 ### Maintenance
 ### Refactoring
+- Added spotless check in the build ([#515](https://github.com/opensearch-project/neural-search/pull/515))